+++ packages/shared/scripts/seeder/load-seed-clickhouse.ts
@@ -1,8 +1,8 @@
 import { randomUUID } from "crypto";
-import { prisma } from "../src/db";
-import { getDisplaySecretKey, hashSecretKey, logger } from "../src/server";
+import { prisma } from "../../src/db";
+import { getDisplaySecretKey, hashSecretKey, logger } from "../../src/server";
 import { prepareClickhouse } from "./prepare-clickhouse";
-import { redis } from "../src/server";
+import { redis } from "../../src/server";
 
 const createRandomProjectId = () => randomUUID().toString();
