+++ packages/shared/scripts/seeder/utils/clickhouse-seed-constants.ts
